pyproject.toml
[project]
name = "ip-design-agent"
version = "0.1.0"
description = "RAG + LangGraph agent for IC physical design and timing analysis"
requires-python = ">=3.11"
dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-postgres>=0.0.12",
    "langgraph>=0.2.0",

    # Embeddings & retrieval
    "openai>=1.50.0",
    "rank-bm25>=0.2.2",

    # Database
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.3.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",

    # MCP
    "mcp[cli]>=1.0.0",

    # UI
    "streamlit>=1.38.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",

    # Document loading
    "pypdf>=4.0.0",
    "unstructured>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ragas>=0.2.0",
    "langsmith>=0.1.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ip_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
